+++ support/python/tests/test_webhook.py
@@ -1,7 +1,9 @@
 import asyncio
-from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Mapping
+import threading
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Mapping, Union
 
 import pytest
+from hypercorn.typing import ASGIFramework
 from nebula_sdk.util import SoftTerminationPolicy
 from nebula_sdk.webhook import WebhookServer
 from quart import Quart
@@ -10,7 +12,7 @@ from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.util.retry import Retry
 
 if TYPE_CHECKING:
-    from wsgiref.types import StartResponse
+    from wsgiref.types import StartResponse, WSGIApplication
 
 
 retry_strategy = Retry(
@@ -23,6 +25,32 @@ session.mount('http://', HTTPAdapter(max_retries=retry_strategy))
 
 
 class TestWebhookServer:
+
+    async def _test_app(self, event_loop: asyncio.AbstractEventLoop,
+                        app: Union[ASGIFramework, 'WSGIApplication']) -> None:
+        # Create the server.
+        term = SoftTerminationPolicy()
+
+        srv = WebhookServer(app, termination_policy=term, port=0)
+        srv_task = event_loop.create_task(srv.serve())
+
+        # Issue request to server and check response.
+        resp = await event_loop.run_in_executor(
+            None, session.get,
+            f'http://localhost:{srv.port}',
+        )
+        resp.raise_for_status()
+
+        assert resp.json() == {'success': True}
+
+        # Stop the server.
+        await term.terminate_task(srv_task)
+
+        # Wait for everything to clean up and exit.
+        await asyncio.wait_for(asyncio.gather(*filter(
+            lambda t: t != asyncio.current_task(),
+            asyncio.all_tasks(),
+        ), return_exceptions=True), 30)
 
     @pytest.mark.asyncio
     async def test_asgi_2(self, event_loop: asyncio.AbstractEventLoop) -> None:
@@ -43,20 +71,7 @@ class TestWebhookServer:
                     'body': b'{"success": true}'
                 })
 
-        term = SoftTerminationPolicy()
-
-        srv = WebhookServer(Application, termination_policy=term, port=0)
-        event_loop.create_task(srv.serve())
-
-        resp = await event_loop.run_in_executor(
-            None, session.get,
-            f'http://localhost:{srv.port}',
-        )
-        resp.raise_for_status()
-
-        assert resp.json() == {'success': True}
-
-        await term.terminate()
+        await self._test_app(event_loop, Application)
 
     @pytest.mark.asyncio
     async def test_asgi_3(self, event_loop: asyncio.AbstractEventLoop) -> None:
@@ -75,20 +90,7 @@ class TestWebhookServer:
                 'body': b'{"success": true}'
             })
 
-        term = SoftTerminationPolicy()
-
-        srv = WebhookServer(application, termination_policy=term, port=0)
-        event_loop.create_task(srv.serve())
-
-        resp = await event_loop.run_in_executor(
-            None, session.get,
-            f'http://localhost:{srv.port}',
-        )
-        resp.raise_for_status()
-
-        assert resp.json() == {'success': True}
-
-        await term.terminate()
+        await self._test_app(event_loop, application)
 
     @pytest.mark.asyncio
     async def test_wsgi(self, event_loop: asyncio.AbstractEventLoop) -> None:
@@ -97,20 +99,7 @@ class TestWebhookServer:
             start_response('200 OK', [])
             yield b'{"success": true}'
 
-        term = SoftTerminationPolicy()
-
-        srv = WebhookServer(application, termination_policy=term, port=0)
-        event_loop.create_task(srv.serve())
-
-        resp = await event_loop.run_in_executor(
-            None, session.get,
-            f'http://localhost:{srv.port}',
-        )
-        resp.raise_for_status()
-
-        assert resp.json() == {'success': True}
-
-        await term.terminate()
+        await self._test_app(event_loop, application)
 
     @pytest.mark.asyncio
     async def test_quart(self, event_loop: asyncio.AbstractEventLoop) -> None:
@@ -120,17 +109,25 @@ class TestWebhookServer:
         async def hello() -> str:
             return '{"success": true}'
 
+        await self._test_app(event_loop, application)
+
+    def test_serve_forever(self) -> None:
+        def application(environ: Mapping[str, Any],
+                        start_response: 'StartResponse') -> Iterable[bytes]:
+            raise NotImplementedError()
+
         term = SoftTerminationPolicy()
 
         srv = WebhookServer(application, termination_policy=term, port=0)
-        event_loop.create_task(srv.serve())
+        t = threading.Thread(target=srv.serve_forever)
+        t.daemon = True
+        t.start()
 
-        resp = await event_loop.run_in_executor(
-            None, session.get,
-            f'http://localhost:{srv.port}',
-        )
-        resp.raise_for_status()
+        # Issue a request to the server; discard response.
+        session.get(f'http://localhost:{srv.port}')
 
-        assert resp.json() == {'success': True}
+        # Request termination of server.
+        term.terminate_all()
 
-        await term.terminate()
+        t.join(timeout=30)
+        assert not t.is_alive()